+++ tests/findex.test.ts
@@ -14,6 +14,7 @@ import {
   UpsertEntries,
   LocationIndexEntry,
   KeywordIndexEntry,
+  generateAliases,
 } from ".."
 import { USERS } from "./data/users"
 import { expect, test } from "vitest"
@@ -438,6 +439,7 @@ async function run(
   upsertEntries: UpsertEntries,
   insertChains: InsertChains,
 ): Promise<void> {
+  console.log("Running!!!")
   const findex = await Findex()
   const masterKey = new FindexKey(randomBytes(32))
   const label = new Label(randomBytes(10))
@@ -503,6 +505,7 @@ async function run(
           ),
           keywords: new Set([Keyword.fromUtf8String("SomeAlias")]),
         },
+        ...generateAliases("SomeAlias"),
       ],
       masterKey,
       label,
@@ -511,19 +514,28 @@ async function run(
       insertChains,
     )
 
-    const results = await findex.search(
-      new Set(["SomeAlias"]),
-      masterKey,
-      label,
-      1000,
-      fetchEntries,
-      fetchChains,
-    )
+    const searchAndCheck = async (keyword: string): Promise<void> => {
+      const results = await findex.search(
+        new Set([keyword]),
+        masterKey,
+        label,
+        1000,
+        fetchEntries,
+        fetchChains,
+      )
 
-    expect(results.length).toEqual(1)
-    expect(results[0]).toEqual(
-      IndexedValue.fromLocation(Location.fromUuid(USERS[0].id)),
-    )
+      expect(results.length).toEqual(1)
+      expect(results[0]).toEqual(
+        IndexedValue.fromLocation(Location.fromUuid(USERS[0].id)),
+      )
+    }
+
+    await searchAndCheck("Som")
+    await searchAndCheck("Some")
+    await searchAndCheck("SomeA")
+    await searchAndCheck("SomeAl")
+    await searchAndCheck("SomeAli")
+    await searchAndCheck("SomeAlia")
   }
 
   {
@@ -560,6 +572,52 @@ async function run(
     expect(results.length).toEqual(100)
   }
 }
+
+test("generateAliases", async () => {
+  {
+    const aliases = generateAliases("Thibaud")
+
+    expect(aliases.length).toEqual(4)
+
+    for (const alias of aliases) {
+      expect(alias.indexedValue).toEqual(
+        IndexedValue.fromNextWord(Keyword.fromUtf8String("Thibaud")),
+      )
+    }
+
+    expect(aliases[0].keywords).toEqual(
+      new Set([Keyword.fromUtf8String("Thi")]),
+    )
+    expect(aliases[1].keywords).toEqual(
+      new Set([Keyword.fromUtf8String("Thib")]),
+    )
+    expect(aliases[2].keywords).toEqual(
+      new Set([Keyword.fromUtf8String("Thiba")]),
+    )
+    expect(aliases[3].keywords).toEqual(
+      new Set([Keyword.fromUtf8String("Thibau")]),
+    )
+  }
+
+  {
+    const aliases = generateAliases("Thibaud", 5)
+
+    expect(aliases.length).toEqual(2)
+
+    for (const alias of aliases) {
+      expect(alias.indexedValue).toEqual(
+        IndexedValue.fromNextWord(Keyword.fromUtf8String("Thibaud")),
+      )
+    }
+
+    expect(aliases[0].keywords).toEqual(
+      new Set([Keyword.fromUtf8String("Thiba")]),
+    )
+    expect(aliases[1].keywords).toEqual(
+      new Set([Keyword.fromUtf8String("Thibau")]),
+    )
+  }
+})
 
 /**
  * @param a one Uint8Array